Adding text files containing algorithms
These text files contain algorithms to sort, populate and draw our tree's. Done in a text file as they require to be part of a larger project to work. They all need adapting to fit our actual project with variable names etc and not everything is done as I don't know how we would be doing certain things in the actual project. Overall though they should be relatively easy to enter into our program as they are largely logic based and will just require some translating. They are written in c++ and should be able to be copy and pasted for the most part.

NOTE - The tree drawing algorithm is not pretty at all but should work if we cant think of a better way to do it.